[make] add -MP to the compile line
This seems to fix a long-standing issue that occurs if a header file
is removed or moved to a new location, causing the make system to
bomb out on the next build due to stale .d file dependencies on the old
location. -MP adds empty rules for all of the .h files to the .d file
which at least lets it get past this stage.