[platform][or1ksim] add or1ksim platform for the or1k architecture
or1ksim is the "golden reference" simulator for the or1k
architecture, but the configuration used in this simple setup
is commonly used on other targets, so it could serve as a good
starting point for those as well.

Included is also a configuration file with uart and ethernet
peripherals defined, which of only the uart is supported.