[platform][stm32f7xx] disable hardware checksumming for ethernet
Turns out if you stuff a valid checksum in an icmp packet, the tx
hardware checksum engine will read it as well when computing its
checksum, and results in a zero stuff.

Doesn't seem to be the case for ip, udp, and tcp.

Disable for now until it's possible to inform the minip stack to not
stuff checksums dynamically.