[pcie] Eliminate IRQ share mode.
Selecting an IRQ sharing mode when selecting an IRQ mode is something
we are not using right now.  Currently, its just a value which you can
get wrong and cause your pcie_set_irq_mode call be rejected (gotcha-APIs).
Additionally, the concept of how IRQs are allocated and shared (and if
they are bound to specific cores, or sub-groups of cores) is more
complicated then the assignment of a simple share mode.

If/when we bring a feature like this back, it will look pretty
different than this API.  For now, just eliminate it as it serves no
purpose.

Change-Id: I5626999ee6739060df1f15b629f0b0e36ec74f75