[build] generate build-*/sysroot/{include,lib,...}
User libraries marked for export with MODULE_EXPORT := name
are packaged up under sysroot/lib/lib<name>.a and their public
headers are packaged under sysroot/include/...

libc is a special case as its crt*.o is separated from the .a
file to be standalone (so it may be linked first) and its headers
are installed under sysroot/include/...

The SYSROOT_MEGA_LIBC option (enabled for now) combines the listed
libraries into one hugenormous libc.a, which is a gross violation
of layering and symbol visibility, but simplifies the link arguments
for the moment.  Once we have shared libraries we can abandon this
horror.

Change-Id: I4d1c9cf379222cfafcaf3025c7243c3fbd1b0fd1