[kernel][magenta] Rough pass at categorizing errors
This organizes the error space into a few categories, removes a few values
that are never or barely ever used, and adds TODOs for several additional
error values that could use additional rationalization. This is a rough pass
to get things in better shape, these error codes or values are by no means
final.

Change-Id: Ic93f33a09b2cfce2621db8d2ce0fb0afd85591cc