[pcie] Soften ASSERTS during init.
Attempt to eliminate any ASSERTs in the pcie_init path which can be
caused by external factors.  If a device's configuration space looks
bogus or illegal, just don't enumerte the device (instead of
ASSERTing).  If the platform gives us an invalid set of resources,
just don't start up instead of melting down.

This CL makes some progress on cleaning up some of the ASSERT vs.
runtime-error-handling behavior in the bus driver, but there is
probably more to be done.

Bug: MG-58
Change-Id: I744b4821f592130378d59535c7f4efb11e034103