[errors] Rename ERR_GENERIC -> ERR_INTERNAL, part 1
This adds ERR_INTERNAL with the same value (-1) as ERR_GENERIC as the
first part of renaming the code. Due to the widepsread use of
ERR_GENERIC and risk of collision, I'll remove the ERR_GENERIC define
in a separate patch so code has a chance to start using the new value
before the old value vanishes.

Change-Id: I15bce2c9930818d6c84d37da80ab7d24de0a7bf6