[kernel][arena][handles] Improve arena.h memory management
.. and up magenta handle limit to 32K handles.

So I didn't realize that I was using page-fault demand paging
(vmm_flags == 0) so the handle limit was low.

So I decided to up the limit and make it so the arena allocator
does not suffer page faults for the data block, since they are
slower.

The control block is still using page faults, but it is usually
far smaller (8 to 32 times) than the data block.

I am not sure if I am operating the vm api correctly though.

Change-Id: Iae7069672bf322d42a48d06d276c6a4d7c406d19