[vmm] Fix behavior for faulting on user data from the kernel
A bug was introduced that would cause a pagefault loop instead of a
proper crash.  This patch can be tested with the
bad-kernel-access-test-crashes command.  Currently only really changes
behavior on x86, but should work on any architecture that will give you
a fault on accessing user data from the kernel domain.

Change-Id: I61f82b09ce802cead62e9feb308c89a91e5cd239