[x86][registers] Rework extended register saving
This patch adds support for using xsave/xsaves to save register state.
It removes support for the optimization previously in place for
reducing FPU state saving, due to the xsaves/xsaveopt support making it
unnecessary.  We could keep the optimization around for older hardware,
but it adds complications to the logic in order to support hardware
that is likely out of our scope of concern.

This patch additionally introduces support for using the YMM and ZMM
registers.

This patch supports the following configurations, in order of
preference:
- xsaves support
- xsaveopt support
- xsave support
- fxsave support only
- No fxsave/xsave support

Change-Id: I8cb4fe6cff00f8a0fe3644de4ab425a22f1cae15