Exception handling version 2.
Major changes:
- exceptions are replied to using the handle of the process of the
  thread that got the exception
- use ioports instead of message pipes
- add "gone" reports so that exception handlers can be notified when
  something they've attached to is gone
  - one *could* fold this into having handlers track SIGNALLED signals
    of the underlying process, thread. for now I have not done so,
    we'll see what falls out once we have some actual user-space
    handlers

Note that there is a bit of code still to deal with "self process"
and the absence of an object representing the "system".
Left for another day.

I had to move the system call to obtain a debug handle of a process
given its koid from the debugger API to this API because exception
report replies now require a process handle which the system exception
handler will not already have when it gets an exception report.

Change-Id: I8805b8b34b7d74d04ba35a1633b8df92cfc1d1d7