Go back to using a mutex in Waiter.
The reason it was a spinlock was so that it (Waiter and Event) could
possibly be used someday from IRQ context (MG-146). After some
discussion (cpu, johngro), we decided that this probably isn't the right
thing, since the Waiter may need to wake up a large number of things (it
iterates over a list under the lock).

Instead, the plan is to use I/O ports from interrupts instead.

Change-Id: I53b217298fb84b893ee082ed21a5c545ccd8e5e2