[utils] Migrate DoublyLinkedList<> users.
Migrate the users of utils::DoublyLinkedList<> to the new API.  Remove
the old implementation and tests.

Note: User definitions of the old list types were preserved as lists
of unmanaged pointer.  This code still needs to be reviewed to see if
it is more appropriate to change it to a list of RefPtr<>s or a list
of unique_ptr<>s

Change-Id: I91005e7ac05f98a843d6f10c29d25d52fdc21782