[process] Don't expose the global process list.
ProcessDispatchers are ref-counted objects in the kernel as of today.
When created and destroyed, they a tracked on a global list of
unmanaged pointers to the Dispatchers.  This is OK, as the lifecycle
of the ProcessDispatchers is not managed by the list itself.  As long
as PDs go on and off the list in a locked context inside of the PD
constructor and destructor, and every operation performed on the list
happens inside of the locked context and does not export an non-ref'ed
copies of any of the pointers, everything should be safe.

Before this CL, however, the list and the lock existed outside of the
ProcessDispatcher class.  Move the list and its lock into the class
as private, static members in order to keep all of the code together
and minimize the chance that someone might do something bad with the
raw pointers which are in this global list.

Change-Id: I7ae775adb348b510ccc1c26e997b07c4bbe6f333