[magenta] Make StateObserver::OnCancel() able to indicate that it wishes to be removed.
I.e., if an obs->OnCancel(handle, should_remove) sets *should_remove to
true (when called by a StateTracker st), then obs will be removed from
st and st.RemoveObserver(obs) should *not* be called in that case.

This will be useful for objects (e.g., wait sets, maybe also I/O ports)
that are "persistent" observers. The trouble is that OnCancel() is
called under the StateTracker's mutex, so it can't call RemoveObserver()
itself. Such objects may not have a convenient place to call
RemoveObserver(), e.g., if they're not currently in a wait.

Conversely, it is not a problem for wait-one/wait-many: The pattern
there is to call AddObserver(), wait (which could be ended from another
thread by OnCancel() by signaling an event, for example), and then call
RemoveObserver().

Change-Id: I69f7432ce0ea0db59c9cd5b6ea2bf8c39581c8c7