[kernel][magenta] IOPort refactoring
Previously pieces of IOPort were directly grafted into the Waiter
object. Now that the Waiter is a StateTracker it is more generic
so we want to refactor things.

1- Remove io-port specific logic out of StateTracker
2- The IOPOrt is now split on the IOPOrt proper and a StateTracker
   observer (IOPortObserver).
3- Add another callback from StateTracker, so we can more cleanly
   manage the lifetime of observers.

Now that the IOPort holds a the list of observers, it paves the way
towards a future where the io port does not have an fixed queue
and therefore it cannot overflow.

Change-Id: Ic9727e8982b1ef2ced061273a11af9dd406a8113