[build] Fix ABI stub generation for undefined symbols
Undefined symbols are in fact a part of the ABI for a DSO, an an
indirect fashion.  The linker uses the presence of an undefined
symbol in a DSO to know that it needs to generate a dynamic
symbol table entry for that symbol when creating an executable
referring to that DSO.  Without this, the symbol will not be
visible to the dynamic linker at runtime to satisfy the DSO's
reference.

Change-Id: Ic457440a70bc55760e027db56ffa8d6d2b1392db