[kernel] [uapp] [userboot] Compensate for nm's lies about Thumb addresses
This makes the entry point address extracted for userboot work as a
Thumb-mode branch target, which covers what the compilers do on ARM
today.  To fix this more cleanly we might need to make the scripts use
readelf or something else that does not molest symbol values in ARM
binaries.  The vDSO entry points don't need adjustment because they
happen to be all ARM-mode code.

Change-Id: If6e3984f95e363a53704a80442a9ee602f60024b