[kernel][magenta] Add handle count to Dispatchers
Now dispatchers have two counts, the traditional reference count
and and now the handle count. Lifetime still works as usual but
now we have knowlege of when the handle count reaches zero.

At zero handles a virtual function is called that dispatchers can
override. This is useful for freeing state that is user-mode facing
or for providing the right semantics.

Change-Id: I3a01b2e050b43a8487c90fac4996a9b3c958d4ae