[magenta][syscalls] normalize handle related errors
Per the error unification and rationalization discussion(s) and
as documented in docs/errors.md:

  ERR_WRONG_TYPE:
    The subject of the operation is the wrong type to perform the operation.
    Example: Attempting a message_read on a thread handle.

  ERR_BAD_HANDLE:
    A specified handle value does not refer to a handle.

This change updates all the magenta syscalls to return the above errors
under the conditions matching their descriptions above.

We disable the global application of crash-on-bad-handles policy for now
as a bunch of userspace code was passing bad handles but not being flagged
because we were treating them as invalid arguments.

Change-Id: I6a34d22d295da8fad744f756cab7417e39a4c901