[kernel][vmm] Change VMM page permission polarity
This makes all page permissions explicit and with positive
polarity.  Previously, specifying no flags would give you
RWX permissions, and you'd need to pass RO to drop the write
permission and NO_EXECUTE to drop the execute permission.

This model more closely matches the one we wish to expose to usermode,
and is less tied to the MMU flags of the arm/x86 implementations.  It
should also make it easier to talk about "mapping" regions with no
permissions in order to reserve parts of an aspace.

Change-Id: I20d99298888155f904fa2d1ae4fee7a72bce3f98