[utils][tests] Smooth out LFSR-as-RNG behavior.
LFSRs (particularly ones whose generators have mostly only high order
bits set) can produce long-ish runs of decreasing values when used as
an RNG which simply samples successive core states in order to produce
their values.  Specifically, whenever there is a run of low order bits
set to zero, the RNG will spit out values each of which is
(prev_val/2) for the length of the run.

The LFSR (given an appropriate generator) will still cycle through all
of its core states, but will have this pattern of runs of decreasing
value which make its behavior as an RNG poor.

Counter this effect by building the next random value by cranking the
LFSR N times (where N is the number of bits in the core state) and
using the low order bits of each core state to assemble the random
value.

Provided that N is a power of 2, we know that the LFSR-as-RNG will
generate a maximal cycle.  Logic is as follows...

1) Each unique core state will generate a unique sequence of bits to
   form a random number.
2) The core is advanced N states per random number generated.
3) The LFSR has 2^N - 1 core states.
4) N is a power of 2 (given)
5) Because (2^N - 1) and N are co-prime, we should still cycle through
   all of the possible core states.
6) Because of #1, the output of the RNG should just be some
   permutation of the LFSR core state sequence.

Change-Id: Ica675349a1c92f29a41feddf0b9e289ec29c13e2