[x86] Remove assert from the kernel linker script
This assertion uses LOADADDR function which is the only construct
not supported by LLD. Supporting this function is not trivial as it
implies the ordering in which the linker script is evaluated which
is not necessarily the same for all linker implementations and it
is unclear if or when it will be supported by LLD as it will
require significant changes to the internal logic.

Furthermore, this assertion is not necessary for the evaluation of
the linker script and this check can be replaced with an external
check (eg using readelf after linking) in the future if needed.

Change-Id: If58cb36bddc0ae5c34ce17280088c0d08c84d690