[kernel][userboot] hold refs for the system vmos constructed from existing pages
This avoids a situation where one of these vmos is closed from user space and
the pages are freed out from the middle of the kernel. That shouldn't be a problem,
except it blows up on the arm architectures due to the way the kernel is mapped.

Added TODO to fix once we get the arm kernels moved out of the Big Memory Map.

Change-Id: I10f6bebdb17db4ad72b4cdfed46e13e97bdf3685