[kernel][ioport] Switch to packet per io operation
Up to this point the ioport only generated packets when the bound
object signals changed, as such it was just another StateObserver

With this change an iopacket is generated with every bound object
event. For example two back-t-back writes on a bound pipe generate
two packets. Thusly, the hack in dispatcher-v2 is removed.

The io port client is no longer a StateObserver, a separate client
class is provided for objects to report events to the ioport if
any.

Currently only message pipe is supported.

Change-Id: Ibbef1910c0d0420af6163ee42757310eb038060a