[kernel][debug] Add debugger agent support
This is the kernel support needed to use
https://fuchsia-review.googlesource.com/#/c/8004

In the context of an arbitrary process, the intended
case is to inject a debugger agent in a process that
needs debugging.

Also we now have a single syscall that goes from
koid to process handle, which would be the only
one that needs to be secured.

Change-Id: I29c4c83fff12d5c9b62c46258a31e31456f72583