[syscalls][socket] return ERR_SHOULD_WAIT instead of ERR_NOT_READY
ERR_SHOULD_WAIT is the equivalent to "would block" in Magenta, even
though we never actually block.  It's the "no data or space yet,
please wait for the condition to change" status.

ERR_NOT_READY implies an unknown state that is expected to transition
to a permanent success or failure state (eg, device not fully online
yet).

Change-Id: Ib3f1d3577afc296c12a86ad3cd30462fa5687718