[magenta][ioport] Add support for sockets
When we reworked the ports to be packet-per-event we
removed support except for message pipe.

This CL adds support for sockets, same as for message
pipes only read and peer-closed are supported.

Additionally Dispatcher is modified with a new
member set_port_client() so we can uniformily treat
pipes and sockets (and in the future others) uniformily.

In addition we now send the byte count of for
the MX_SIGNAL_READABLE packet.

Change-Id: Ie510ff677c9623fe95abd97fd2880d3d3a8f1adf