[kernel][debug] beginnings of register r/w support
This provides two syscalls mx_thread_read_state,mx_thread_write_state
that can be used to r/w cpu registers.
It is built off of the idea of "register sets" (integer, floating point,
and so on) in common use at least in gdb.

Discussion of the ultimate name of these syscalls, and whether to
merge them with something else is being left until the raft of
dependencies relying on this code is ready.

Change-Id: Ic118a76292334dfeaa46278aaddca3140ed4441c