[kernel][usr_ptr] convert low level VMO Read/Write routines to using usr_ptr<>
This has rippling changes out to any code that touches this api, and generally
converts more syscalls to using the usr_ptr wrapper class.

There's still more to do, but this is enough for the parts that eventually get
to vmos.

Also move user_ptr.h out of mxtl into lib/user_copy in the kernel.

Change-Id: I570254706a060562b262ede0a2e704219e168eb0