[magenta][signals] Accumulate wakeup reasons.
Record all of the reasons that a WaitStateObserver might become
signalled while it is on the list of observers for a given
StateTracker and return that as the reason that a thread was woken
during a wait one or wait many operation.

Previously, when a thread unblocked during a wait one or wait many
operation, it would report the current set of satisfied and
satisfiable signals in the StateTracker as its reason for waking up.
This meant that it was technically possible to assert and de-assert a
signal quickly enough that a thread which was waiting for signal X
might wake up, but be told that signal X was not currently asserted.
This can make it difficult to determine the reason why a thread was
unblocked; a problem which becomes even more complicated when a thread
is waiting for multiple signals on multiple handles simultaneously.

Now WaitStateObservers store two bitmasks internally, one for the
satisfied signal state and one for the satisfiable signal state.  When
the observer joins a StateTracker, the bitmasks are initialized to the
current signal state of the tracker.  Anytime a change is made to the
signal state of the tracker which might wake up an observer, the
changes are accumulated in the bitmasks.  In particlar, signaled
conditions are OR'ed into the signal mask (as a signal becoming set
can wake up an observer) and satisfiable conditions are AND'ed into
the satisfiable mask (as a signal becoming unsatisfiable can wake up
an observer).

The result is that users are told which of the possible conditions
which might wake them held at any time during their observation,
instead of capturing the conditions at the instant they left the
observer set.

Change-Id: I3286d379271192cbefce3b1a91afbe96eea085e9