[kernel][c++] move the global constructor calls to very early in the kernel initialization
Now it's called more or less immediately upon entering the kernel, before
almost all of the subsystems. This means you have to be more careful
what you do in any global constructors, but they're pretty much only
being used to initialize various list and container objects, so it's
not a big problem.

Change-Id: Id7c3e12d85457f99ceb2e24359e1b9ec48c571ba