[kernel][io] add some slight safety checks to the serial output path
Unproven that this is a problem, but if the thread structure got badly
corrupted, the serial output routine can easily go into a loop that
spews garbage and locks up any other cpus that try to serial output.

At least try to catch it so that we dont end up in a endless spew.

Change-Id: I4a1c08a2573c3c50ed210ea9eacec1b709279d6b