x86-64: Stop using the inherited stack when paging is enabled.
The stack might not be in the first GB of memory, which is the only part which
is still identity mapped once paging is initially turned on. We have to
continue to use the physical address because the virtual address is 64 bit and
is not accessible to us in compatibility mode, the mode we're in briefly after
paging is enabled but before we switch fully into 64 bit mode.

Change-Id: Id776e9d0f184f68a5f73294c65a0d2d094135e40