[kernel][x86] Fix asm operand constraints in %gs:N accessors
The "g" constraint tells the compiler that an immediate, register,
or memory reference are all acceptable in the instruction.  That's
not true, since in these instructions there is another operand
(the %gs-prefixed one) that is a memory reference, and the x86 mov
instructions do not allow two memory-reference operands.  The "ir"
constraint says what we mean: best operand is immediate operand;
acceptable operand can also be register operand.  For the output
operand cases, "=r" is the only valid constraint: the output operand
of the mov instruction must be a register (since the input operand
is a memory reference).

Change-Id: I32545941715f52e776cfce907d04580717020e2f