[kernel][pmm] Don't give up when the first physical memory region is exhausted.
As far as I can see nothing removes an empty physical memory arenas from the
list. If such an arena is run across when in pmm_alloc_pages, the function
will stop looking for pages and return the number it's found thus far. That
may not be the number requested, and may even be zero if there was a totally
exhausted arena (unless I'm missing something).

This change prevents a "blue screen" when booting with depthcharge and OVMF on
QEMU.

Change-Id: I042827c3859c49c59885c58c363b4de00f74107d