[pc][timer] Improve calibration routines
This reworks timer calibration to perform two groups of runs
for different durations.  In each group of runs, the fastest
run is kept, since it's not possible for the code to run faster
than real time, and there is potential for large anomolies in runs
due to things like SMIs.  We subtract the two kept values, and divide
by the difference in durations, with the idea that the overhead
in measuring is constant.  We end up with
measured_i = real_i + overhead, so
measured_1 - measured_0 = real_1 - real_0, and we can use our
reference timer to determine the tick rate.

This mechanism appears to provide an order of magnitude improvement
in accuracy.  The accuracy still is not great (my tests currently
show ~700ppm in the worst of boots), it is better than the 5000ppm
we were seeing before.

Change-Id: Ib2fede6a217b0a28b174e55a412cb1ca98a9810d