[x86][mmu] Fix bug related to unmapping large demand paged regions
If a demand-paged region of size >=2MB was created, and it was not
aligned to 2MB, and it straddle a 1GB boundary, crashing without
mapping any pages into it would cause misaccounting to occur in
the MMU code, resulting in an assert firing.

The same misaccounting could cause pages to not be unmapped when
requested.

MG-315 #done

Change-Id: I7526eb370b4aba3671c87d78681018f62ba95e26