[gfxconsole] Let the early-init console run for longer.
Now that PCI leaves BIOS/bootloader configurations in place, don't
shutdown the early init console so early in the boot process.  Instead
let it run all the way through PCI initialization.

Note: There should be a better way of handing off kernel graphics to
user-mode, but right now PCI init is a pretty good proxy for when
user-mode is going to take over graphics.

Also, change the default background color for the kernel graphics
console to be opaque black instead of transparent.  This only matters
for framebuffers running in ARGB mode.

Change-Id: I6c0d1f24754db7d8e46a322b6618d20d63e24b4f