[kernel][userboot] New VmObject::CreateFromROData static method
This moves the code to create a VMO and populate it from kernel
RO-data pages into a static method on class VmObject.  The
knowledge of the VM layer's limitation about reusing pages from
the kernel image mapping is moved into this VmObject method,
rather than being implicit in the userboot code.

Change-Id: I6dce9b4e72afde7c530f309a79b8c5db5702a68e