[kernel][magenta] Job objects (part 2)
In this installment we 'thread' the root job which is held in a
global into the userboot process, which has a handle to it.

The userboot sends the job object to the device manager which
then creates a child job object, which is not used at the moment.

The change includes a syscall to create a job object which only
devmanager can call since its the only one that gets passed a
job object.

the `mx ps shoud show at total of two job object now, like this

[00017.656] K       id-s  #t  #h:  #jb #pr #th #vm #mp #ev #ip #dp [name]
[00017.656] K      270-R   6 104:   2   1   6   4  84   0   3   0  [bin/devmgr]

Change-Id: If406588228e8b9dce641975add5194d7cfdc33b1