[kernel][arm64] Fix operand sizes in inline asm
The code is using operand formats like %0 that mean the 64-bit
register, but operand values that are 32 bits.  These are supposed to
match, but GCC does not check that they do (Clang does).  In assembly
operands where the 32-bit register is valid, the operand format can be
%w0 to emit the 32-bit register name instead of the 64-bit register
name.  But for instructions like mrs, msr, and tlbi, the register
operand must be a 64-bit register.  So the existing code was emitting
assembly like "msr x0, fpcr" but preparing the value using w0.  This
is probably harmless in all cases because using wN clears the high
bits of the register.  But it's sloppy, and calculating in 64-bit
registers doesn't cost any more.

Change-Id: If55248b0bdbd44c4a95401e26b1e1b7a21789d54