[pcie] Rip out kernel mode pcie drivers.
Get rid of the concept of kernel mode pcie drivers.  Start to lay the
groundwork for re-defining the concept of "claimed" in order to allow
user mode code to hold multiple references to a PCI device, at most
one of which may have the device claimed at any point in time.

Change-Id: I7596de831f40feca6f846c5f78d5f9283a673abc