[pcie] Refactor to add PcieRoot
Refactor PcieBridge so we can model PCIe roots as objects which manage a bus and
have children, but are not devices.  Specifically, there is now a
PcieUpstreamNode class which manages children, and is a base class for PcieRoot
and PcieDevice.  "Having children" is implemented by UpstreamNode, "Being a
device with children" is implemented by PcieBridge, and "Being a non-device with
children" is implemented by PcieRoot.

Next steps
++ Get rid of the root_complex_ singleton member of the bus driver.  Replace it
   with a collection roots.
++ Move the management of legacy IRQ swizzling out of the PciePlatform class and
   into the PcieRoot class.  Swizzle behavior is platform specific, but also
   specific to the traversal of a specific root.  When the platform adds roots
   to the driver, the swizzle behavior of that specific root should be provided.

Change-Id: I62118429632eb30a827f50b49df737eda0f4f693