[futex] Make the interface of RemoveFromHead() easier to understand
It wasn't obvious that RemoveFromHead() always removes at least one
node.  Make this clearer by commenting it and adding assertions for
the two preconditions that make this property true.  RemoveFromHead()
was never being called with count==0; the callers all check for this.

Make RemoveFromHead() a static function (instead of a method) so that
it's easier to refer to the list_head argument and explain its
properties.

Note that the variable "last" does not need to be initialised given
the preconditions, so I've replaced its initialiser with nullptr.

This is a cleanup to help with MG-354.

Change-Id: I90147699dde4509a61e322fe41f2b197ba114880