[kernel][mangenta] enable object_signal() for all objects
With this CL mx_object_signal() works for all objects that
have a state tracker (so get_state_tracker() != null) and that
not explicitly opt-out.

This provides an easy way to break out from a wait by simply
waiting for MX_USER_SIGNAL_0 in addition to whatever they are
interested on.

This is useful for:
- job
- process
- thread

and was already true for:
- event

This method of wait breaking is usually superior to wait
cancelling (via closing a waited handle) _if_ the construct
is a message loop _and_ the loop is being continued.

Currently there is special behavior on object_signal() for
some objects in which they signal the other end rather than
the local end. So wait breaking does not work.

This is true for:
- socket
- event_pair

But not true for
- channel

This needs to be rationalized in a follow up CL.

Change-Id: I7a805de6458b3b59ea63bfe2fde9ff186fbb3204