[futex] Reduce internal lock contention between FutexWait()/FutexWake()
FutexWake() needs to be holding FutexContext's internal lock when it
calls cond_signal() to wake the target thread.  Unfortunately, when
the target thread wakes up, cond_wait_timeout() tries to claim that
lock, causing contention -- the target thread will often block again.

There is a simple way to fix this.  Although a condvar wait operation
must be called with the condvar's mutex held (so that the release+wait
is atomic), it is not essential for the condvar wait operation to
reclaim the mutex before returning.  Reclaiming the mutex (if it's
done at all) does not need to be combined atomically with waiting.

So, we can add a non-mutex-reclaiming variant of cond_wait_timeout()
and use that in this case.

This is a minimally invasive way to reduce contention, which preserves
the existing layering in which wait_queue is only used inside
kernel/kernel/.  We could improve on this by using wait_queue
directly, or by reducing the overhead of mutex_t, using
per-hash-bucket locks, etc.

MG-354

Change-Id: I67dfd46a89fc681f2b8df24169e771fd4bb456b4