[kernel][threads] Remove process name from thread name.
If a thread name is gettable and settable via a property,
then storing the process name with the thread name is problematic.

This keeps debugging printfs with thread names as is.
That means they will lose the process name in the output.
I can imagine this being inconsequential for most cases.
OTOH I can imagine wanting to keep the process name in the
output for the context switching printfs.
Leaving this to be fixed at a later time if/when needed.

Change-Id: Ia404e25e7f5587014af7d99cfd9bf8af7a9a8e47