[kernel][arena] up front allocate memory for the arena
Simplifies the logic a bit and avoids fatal kernel OOM conditions
where it runs out of memory as it demand faults in a page for an arena.

For now, just up front allocate all of the pages for the arena so that
it's unaffected by OOM.

On a 64bit machine with 256k handles, this equals about 14MB of memory
usage.

Change-Id: I21020b4e5ba8c00d1f5d3973ea921bfbdb17dd74