[state_tracker] Adding an observer cannot fail
This used to be able to fail, when this class was polymorphic, with
behavior dependent upon whether or not it was to be called from an IRQ
context. c93d682d811110d5ac68b6d0ef1965f565f725fe removed this
complication. Not returning an always-NO_ERORR status simplifies some
syscall code, and in particular and makes it easier to see which error
codes can be returned to userspace.

Change-Id: I1f39e2aeb27b31b3bd42f123e177646da01c2a2d