[syscalls] switch to generated code
...from the macro and syscall.inc universe

Most of the turbulence is changing the badgercast back
to T* and the adding user_ptr<T> in each spot which
looks less pretty but it is correct by construction
and a temporary state anyhow.

In the final state the syscalls would look and speak
pretty again, but that needs more codegen.

Along with this we also change pointer/buffers from
user mode to have underscore prefix, so we are consistent
there.

Change-Id: I8342210f667f5056ac56a5bab168cb8776ab35ce