[system][ulib] mini-process and [kernel][job] kill
Main new functionality is that mx_task_kill() works with jobs.

That however uncovered a bug with with a new test.
The bug was in using the wrong trait in JobDispatcher::RemoveChildProcess()
which mean the wrong double linked list was modified.

instead of process.dll_process_ we needed to use process.dll_job_

But in order to write the test we need a new user library: mini-process
which allows you to spawn a functional process without a new binary.
Such is needed for core-tests which assume launchpad is not available.

Change-Id: Ie16a94c8a8bd91bd28f3c0071c8a4cbff9573e59