[magenta] Remove Dispatch::get_specific() and solidify DownCastDispatcher
get_specific() returned a raw pointer to a reference counted object
which is tricky to use correctly without leaks, as evidenced by the
leak in DownCastDispatcher. With our RefPtr and C++ move semantics
there is no need to escape to raw pointers - it is fine and preferable
to remain in RefPtr<> objects and move when appropriate to avoid
reference counting churn. This implements a non-leaking version of
DownCastDispatcher and transforms calls of the form

  mxtl::RefPtr<Dispatcher> dispatcher = ....;
  DispatcherSubclass* disp_subclass =
      dispatcher.get_specific<DispatcherSubclass>();

(with disp_subclass normally hidden in 'auto') to this:

  mxtl::RefPtr<Dispatcher> dispatcher = ....;
  mxtl::RefPtr<DispatcherSubclass> disp_subclass =
      DownCastDispatcher<DispatcherSubclass>(&dispatcher);

This moves the reference count into the subclass object when the type
check succeeds.

Change-Id: I950a46273bd513d7c178016a7d222ca59669cd9b