[arm][user_copy] Fix handling of kernel addresses
As a comment in the code notes, the ARM unprivileged read/write
instructions do not cause a fault that distinguishes between
the fault being from a privileged or unprivileged page access.
As a work-around, we check if the address is in the user half
of the address space before attempting to copy to it, since
we never map privileged pages into the user address space.

Change-Id: If33f53dd672e46eeefc91f41c448903743623448