[magenta][syscalls] implement mx_channel_call()
This is a fused write-wait-read syscall for the use of synchronous rpc
callers.  The kernel treats the first uint32_t of the outbound message
as a transaction ID and accepts an inbound message with a matching txid
as a reply, completing the transaction.

This allows multi-threaded calling of synchronous rpc transports
with no userspace bookkeeping required and will provide an opportunity
for smarter scheduling of such transactions in the future.

Change-Id: Ifbb92b36618ebf13a6e52c650d16fed9f5071c3a