[kernel][magenta] Add signals to Job objects, reform code
There are now two signals:

- MX_JOB_NO_PROCESSES
- MX_JOB_NO_JOBS

You can wait on them.

So now you can wait until all processes or all jobs are dead.

Along the way we also add a lifetime state to the job object, this
way we can issue mx_task_kill(job) and be sure that if it completes
there are no new processes or jobs that raced their way in.

This results in a new error condition if a process or job is
created when the killing operation is in progress.

In addition, reform the way kill happens by only taking the lock
once, swaping the lists and killing each child.

Change-Id: Id72392ec9282aec51e8401acccb501fa742ca138