Add thread annotation macros and enable Wthread-safety
This defines some thread safety annotation macros, adds a bare
minimum set of annotations to our top-level lock primitives and
helpers, and enables -Wthread-safety when building using clang.

There are actually two sets of macros defined. One is defined in
system/public/magenta/compiler.h with long, ugly names for use in
interfaces exposed in the sysroot where we must avoid colliding with any
identifiers users of our sysroot may use or want to use. These should
only be spelled out in common library code like mxtl.

The second are terser TA_xxx() macros defined in
system/private/magenta/thread_annotations.h that are appropriate for
using anywhere in the Magenta tree that is not exposed through the
sysroot.

There are also a small set of annotation fixes for futex to satisfy the
analysis. These can be tightened up with more extensive analysis
throughout the code.

Our definition and usage of these macros is documented in
docs/thread_annotations.md.

Change-Id: Idd3c08fa277cfdc580c846fe71253100353b2d4b