[magenta][hypervisor] Add guest context to setup VMCS regions.
This is used to set the context from which a guest will execute. This currently
does not work correctly, but we have got to a point where we can vmlaunch
successfully. From here, we need figure out what to setup to enter a guest, and
to exit back to the host.

Change-Id: I340b863223b73b133a7b6e876f2de9f3bb3155ca