[magenta][errors] Rename ERR_REMOTE_CLOSED to ERR_PEER_CLOSED
This error, at least at the Magenta syscall level, is really about
peers. This new name aligns with the MX_*_PEER_CLOSED signals. And
REMOTE implies a distance or asymmetry that isn't necessarily true.

This leaves a compatibility alias for the old name in place until the
rest of Fuchsia can be cleaned up.

Change-Id: Icf95003b2813270d865c6876b62e21d9f5b8c0f6