[magenta][hypervisor] Add logic for host save and load.
With this, we can enter and exit from a guest, and see the reason for
VM exit.  We save the host's state right before we enter into the
guest, and return exactly to that point when exit from the guest. The
return value of vmx_host_save is used to determine whether we are
initially saving host state, or if we are loading the host state back
after a VM exit.

Change-Id: Ia15148b43b131b418fbc69b01d59b3e32945f724