[kernel][hypervisor] Reuse the same dispatcher, if one exists.
Make the hypervisor dispatcher a singleton, so that the processor
being in VMX mode is controlled by the lifetime of the dispatcher.

This prevents multiple dispatcher with different lifetimes disabling
VMX mode whilst it is still in use. For now, the dispatcher's ref
count is also bumped up, so it will never be destructed.

Change-Id: I44c08b336a8b1bc80629ec62546c47d5f641f2eb