[syscalls][port] Guard against invalid options bits in port_create
The port_create syscall passes on options bits to the port dispatch
create functions. The v2 version asserts that it is passed options
containing exactly the MX_PORT_OPT_V2 bit. This should make sense, as
the general pattern is for the syscall layer to bounce back with
ERR_INVALID_ARGS if any unknown bits are set. While this may or may
not be the pattern we eventually want, for now we should be internally
consistent. As is, userspace can trivially trigger the DEBUG_ASSERT.

Part of MG-570

Change-Id: I6d58a6e4b4b488ba3748e6455d8b158f9a1a871d