[kernel][syscalls] Add MX_PROP_PROCESS_DEBUG_ADDR
When aslr is turned on the address of the dynamic linker is no longer
fixed and thus tools like crashlogger and debuggers can no longer find
the address of ld.so's _dl_debug_addr. This patch adds
MX_PROP_PROCESS_DEBUG_ADDR as a way for tools to obtain this value.

MG-536

Change-Id: Icd5b144a68d4d68b4147450fdc8e7371836a8eb6