[kernel][magenta] PortsV2: change dequeue count
For signals in the packet.

Previously when a signals packet was dequeued, the signals.count
contained the number of times the signal had happended since the
last packet had been dequeued. For example when watching
a channel the other side can observe "3 reads" in a single packet.

With this change the count is always 1, so instead the caller will
dequeue 3 packets (in the example above).

Every time a packet is syntetized, internally the queued packet
is put back in the queue until the count reaches zero.

Change-Id: Idc5e19187002755fd64cbfce4ac54072bd17a0f8