[kernel][magenta] Add initial counts to StateObserver
And wire them to the PortObserver.

The main change is in StateTracker::AddObserver() which
now takes an optional CountInfo struct that can have two
generic signal + count entries.

The root dispatcher object grows a new virtual call
add_observer(StateObserver*) which does the right thing
for most objects and its only overriden by ChannelDispatcher

This is a marked improvement from portsv1 which required
a separate PortClient per dispatcher for this same feature.

Note: there is a O(n) there in messages_.size_slow() which
can be fixed. A similiar one existed in the portv1 code
as well.

Test of the feature is done in the portsv2 code.

Change-Id: I31acd276092e0cdce4e4607685c57e89f8728125