[multiboot][bootdata] support non-zeropage boot path
This functionality is gated behind ENABLE_NEW_BOOT
which is not defined in the default build.

When enabled, the Magenta kernel image is a multiboot +
bootdata image and does not use the legacy linux zero
page data passing mechanism.

Since Qemu passes the multiboot commandline and
modules (including the ramdisk) immediately after
the last page of the BSS, we need to carefully avoid
clobbering that with boot_alloc_mem().  To do so
arch/x86's mtrr init now uses a static data structure
(removing the only boot alloc that happens prior to
platform init), and the boot allocator has a simple
boot_alloc_reserve() which pc platform init uses to
inform it of the location of the ramdisk so that it
can shift its "heap" after it if necessary.

After pmm init, pmm_alloc_range() is used to preserve
the ramdisk, the same way we do in zero page boot.

Kernel is now built at 1MB instead of 1MB - 8KB
thanks to changes from travisg.

Change-Id: Iff5a7e1e5dd3133ca93ffb6ee3159c38fe910089