[kernel][magenta] Fix ReadState,WriteState
The presence of thread_.state != THREAD_BLOCKED is a vestige of
an earlier test. Now that the thread can be running while the
debugger is trying to read state (in the new window introduced by
the patch for MG-504, between the two locks of exception_wait_lock_ in
ExceptionHandlerExchange) this test is no longer innocuous but harmful.

I hope this fixes all of MG-625.

MG-625 #done

Change-Id: I3b07148062d3f97119a88e903a880a036683bfc2