[vmo] Move cache policy into VmObjectPhysical
This change adds cache policy to physical vmos as a temporary measure
while the driver and resource models work themselves out. Physical
vmos by default are configured to map as uncached and this can be
changed via new methods and syscalls. The cache policy can only be
changed if the vmo is not mapped presently, but at this time nothing
prevents two processes with the same unmapped vmo from changing the
policy out from under one another. As it stands, this serves to enable
drivers to move to a resource model with those caveats in mind until
a better solution can be added in the future.

Change-Id: I1a7a5776c5d08b74415f85ba7ac6dfb90c0d2d56