[magenta] enable new boot path by default
This causes magenta.bin to support booting via the Magenta Bootdata
or the MultiBoot v0.97 protocols.  The old boot path was a hybrid
of Linux Zeropage + Magenta Extensions and MultiBoot 0.97, but the
latter was not capable of booting us with an external bootdata image.

Qemu already supports the MultiBoot path, and our local branch has
been patched to speed up ramdisk loading.

GigaBoot 0.5.1 (current version in Magenta) also supports the
new boot path.

If all goes well, the old boot path will be removed in a week or two.

Change-Id: I2a62282565c611f7152f84d2dfd4de870eea0036