[vmo] Fix Physical GetPageLocked
While not ideal, we still would like GetPageLocked to handle on-demand
paging for VmObjectPhysical. This removes some outdated asserts and
checks that resulted in a DEBUG_ASSERT or error with the handling of
_page. The correct behavior is to just return the corresponding paddr_t.
This will allow physical VMOs without mappings to work, but in a general
case they will want to map the range of their memory beforehand.

Change-Id: I18a0a23501691a8e8183adc3e65f872b67dd20d6