[vmar] Improve non-COMPACT randomized allocator
This replaces the pick-a-spot at random allocator with one that
is aware of the existing allocations, so it will never try to place
on top of an existing one.

The old allocator could get unlucky and spuriously fail allocations, and
had some undesirable rounding behavior.

The changed test corrects flakiness due to randomization.

Change-Id: I98550c51c00495601a627effe1c8807d6a56f93e