[kernel][arm64] Don't allocate SMP_MAX_CPUS times the boot CPU stack size; clean up bss
The preallocated kernel stack is used only for the stack of the
boot CPU's initial kernel thread.  It doesn't need to have space
for a stack for each CPU.  The symbols for this stack block also
don't need to be global or have obscure names.

Only the kernel translation table needs to be in prebss.
Put the initial kernel stack in normal .bss instead, and
don't point to it until after .bss is cleared.

Optimize .bss clearing to use 16-byte stores.

Change-Id: Id6cab05f95262bcde0a0b543e1f2952e1e31fb64