[kernel][dev][intel_rng] Use compiler intrinsics
The compilers provide standard intrinsics for the rdrand and
rdseed instructions.  Use them instead of inline asm.  Use of
these intrinsics was previously disabled because of a code
generation bug in GCC-6.2.0 (still present in 6.3.0 as released).
The new GCC-6.3.0 toolchain includes the backported upstream fix
for that bug, so use of the intrinsics should now be safe (and
always was in the Clang build).

Change-Id: Id8c1e38bf454a91d4adec389747362979fcda964