[kernel][magenta] make MX_WAIT_ASYNC_REPEATING more usable
This CL greatly tweaks repeating mode to align very closely
to one-shot mode. The mental mode to apply is that
repeating mode is like the one-shot mode except that it
re-arms automatically when you pull out the existing packet.

As a result some tests are actually mode agnostic but
we run them in both modes because we'r'paranoid.

Change-Id: I765446c8f20d20715115cf04aff7b721aa8e0386