[pci] Add Kaby Lake DID to TOLUD quirk.
Note: I'm still waiting to hear back from Intel about whether or not
this DID is the proper one.  Their docs say that it should be the same
as Skylake (0x19xx), but all of the hardware I have run into and seen
lspci dumps online seem to agree that the proper value for Kaby Lake
is 0x59xx.

MG-684 #done

Change-Id: I0ec00cc85d173effc04091d5e6bc8c4a3c484d66