[kernel][vmo] Workaround a deadlock between vmo_syscalls and pagefaults
This avoids MG-730, which has been causing issues on the buildbots.  It
is not a real fix, but it prevents page faults from happening while
holding the VMO lock during a faulting vmo_read/vmo_write.  This will
stop working once we start paging things out in the future.

MG-730 #comment Workaround

Change-Id: I7ade9fce90f5cb20295f045da425cb34f9f2b5e7