[kernel][magenta] Implement policy guts
.. and wire to one syscall.

This CL shows the end-to-end process of
computing policy and applying it to a
single case: mx_channel_create().

It also changes how policy is declared
making it simpler to process: now
regardless of the condition, the policy
have a single set of constants
MX_ACTION_POL_DENY, MX_POL_ACTION_ALLOW
and so on.

Change-Id: I2b09ce54af3c1ab3332e7b2ab160cd5b768e8dd5