[rpi3][platform] Implement platform_halt_cpu.
The BCM2837 on the RPi3 does not implement a PSCI interface so we
must implement platform_halt_cpu ourselves.

To halt the CPU, we mimic the behaviour of the firmware and park the
CPU in a loop at a low memory address until somebody sets our CPU
vector and issues an SEV instruction.

Change-Id: I91f21269cc3dd224884352903540b9c37563f270