[hypervisor] Correctly handle LAPIC.
I incorrectly set the APIC access address to be the guest physical
address of the APIC access page, when it should be the host physical
address.

With that change, we now need to handle APIC access VM exits, so this
CL provides the basic handling required to get to userboot.

Following CLs will add correct handling of various timer interactions,
and of interrupt injection.

Change-Id: I323170c33bfff500c9cf255c6e88387a1706c0ae