[build] Improve MDI dependencies to avoid full rebuilds
The MDI build rules generate both a MDI binary blob for inclusion
in the bootdata and a C header file called mdi-defs.h that is included in code
that needs definitions for interpreting the MDI data.
mdi-defs.h is added to TARGET_MODDEPS so we can ensure that it will exist
before compiling any code that depends on it.
As a result, regenerating this header file resulted in a full rebuild of all magenta.

In this change we adjust the dependencies for the rule that generates mdi-defs.h
to depend only on MDI_INCLUDES and no longer depend on MDI_SRCS.
Only the MDI source in MDI_INCLUDES should contain definitions that affect the
contents of mdi-defs.h.
Now changing files in MDI_INCLUDES will force recompiling all the C/C++ code

Change-Id: Iff2a99afdc9866493d94148cee5143272ca65e4e