[hypervisor] Add op to trap on guest physical address.
This allows us to handle memory mapped devices in user-space. We can
now trap on access to an address range, and let user-space respond to
the trap. This includes access to unmapped address ranges.

For now, we are simply faulting on unmapped address access. In follow
up CLs, I'll begin the migration of devices to user-space.

Change-Id: I60fc2557e5af5fd284a86c2cd1f6a67db89144e9