[hypervisor] Add basic FADT and DSDT to ulib.
This allows us to start acpisvc. I'll need to set several more things,
and move IO port handling completely to user space, but I'll do that
in a follow up CL.

For now, we're not using the IASL compiler, as there is some
contention on whether we want the dependencies introduced by it.

This CL also makes sure we set RAX to 0 when there is an IO port read
we don't understand.

Change-Id: I2687bb837c3d0538194df01a0a8ef94fb93eed44