[kernel] Fix bug where killing a suspended thread causes it to resume
Before this change, killing a suspended thread causes it to resume and
execute more userland instructions before it is finally terminated.

Fix that by adding an extra check.

There was a test for this, test_kill_suspended(), but it wasn't
effective.  This used test_wait_thread_fn() as the test thread.  The
test only detects the bug if test_wait_thread_fn()'s call to
mx_object_signal() is executed, but the thread never gets that far.
The test thread is suspended either inside mx_object_wait_one() or
before.  Killing the thread would cause the thread to resume and retry
the mx_object_wait_one() syscall.  The syscall internally checks for
THREAD_SIGNAL_KILL and exits, so the thread doesn't get as far as the
mx_object_signal() syscall.

Replace test_kill_suspended() with another test which does detect the
bug.  This uses a thread that doesn't call any syscalls
(test_writing_thread_fn()).

Change-Id: Ifd85199c3d8ed8b17b6e1ae09db89fd1b9c27955