[ulib][magenta] Define DATA_CONSTANTS symbol in assembly
If an LTO compiler can see a const object's initializer, it will
constant-fold references to the const object.  However, the vDSO's
DATA_CONSTANTS object is defined as const but is not really a
compile-time constant.  So define DATA_CONSTANTS in assembly rather
than in C++, making the initialized values invisible to the compiler.

Change-Id: Ic12e0bfb4cc23a51f7d59b0bf8bd668d761c5d6d