[mp] Atomic in_int_handler for arm64
Previously there was a race condition in the arch_in_int_handler
code which could cause the function to return the wrong value.

This was causing asserts to fail (ultimately resulting in kernel
panics).
The solution is to disable interrupts upon entry into the function
and to re-enable them on the way out thereby making the call atomic.

Change-Id: I180f67bc418e6edb1d13f468375fcddd01145b56