[kernel] Add comment to explain thread safety of register state
It wasn't obvious to me why it is safe for functions such as
x86_set_suspended_general_regs() to be setting
thread->suspended_general_regs etc. without taking a lock, in the
presence of accesses to those data structures by
UserThread::ReadState() and WriteState().

Add a comment that should make that more obvious.

Change-Id: I17aa718127b18453346b6c1256592ef14220b345